Update target-triple unit test using new APIs
